Fix build error on x64

+++ EmulatorPkg/Hii2RedfishBootDxe/Hii2RedfishBootDxe.c
@@ -208,7 +208,7 @@ RefreshBootOrderList (
       continue;
     }
 
-    mBootOptionsVarData.BootOptionOrder[OptionIndex++] = BootOption[Index].OptionNumber;
+    mBootOptionsVarData.BootOptionOrder[OptionIndex++] = (UINT32)BootOption[Index].OptionNumber;
 
     ASSERT (BootOption[Index].Description != NULL);
